% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLT.R
\name{CLT}
\alias{CLT}
\title{CLT function}
\usage{
CLT(y)
}
\arguments{
\item{y}{a matrix where every column is a vector of random variables with the same parametervalues.}
}
\value{
Returns a histrogram with the CLT value \eqn{(sum_{i=1}^n Y_i)\sqrt(n)} for each column.
}
\description{
The function that checks, if there is a Central Limit Theorem.
}
\details{
If there is a is a Central Limit Theorem \eqn{(sum_{i=1}^n Y_i)\sqrt(n)} would converge towards i normal distribution. 
If the histogram produced by this function resembles a normal distribution, it would seem, that there is a Cental Limit Theorem.
}
\examples{
y<-simAR1(0,1000,0.8,rnorm(1000))
CLT(y)
}
