% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.simulation}
\alias{plot.simulation}
\title{Plot simulations}
\usage{
\method{plot}{simulation}(p)
}
\arguments{
\item{p}{a matrix of class "simulation".}
}
\value{
Returns a plot of every column of p.
}
\description{
The plot method for objects of class \eqn{"simulation"}.
}
\examples{
p <- rAR1(0, 300, 0.8, distfun = rnorm, 300)
plot(p)
 
}
