% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rAR1.R
\name{rAR1}
\alias{rAR1}
\title{The rAR1 function}
\usage{
rAR1(n, y0 = 0, m = 300, phi = 0.5, distfun = rnorm, ...)
}
\arguments{
\item{n}{nubmer of columns of AR(1) simulations the function shold give.}

\item{y0}{initial value}

\item{m}{numer of steps}

\item{phi}{parameter of the model}

\item{distfun}{a distribution function}

\item{...}{parameters to the function disfun}
}
\value{
a matrix where every column has m steps of an AR(1) process, and every process has the same parameters.
}
\description{
The rAR1 function
}
\examples{
l<-rAR1(10,5,200,0.6,rnorm,200)
summary(l)
plot(l)
}
