% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAR1.R
\name{simAR1}
\alias{simAR1}
\title{AR(1) simulation}
\usage{
simAR1(y0, n, phi, distr)
}
\arguments{
\item{y0}{initial value}

\item{n}{numer of steps}

\item{phi}{parameter of the model}

\item{distr}{vector with n elements with a given distribution.}
}
\value{
returns the four plots described.
y<-simAR1(0,1000,0.8,rnorm(1000))
x<-simAR1(10,300,0,rt(300,5))
}
\description{
AR(1) simulation
}
