% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.simulation}
\alias{summary.simulation}
\title{Object summary}
\usage{
\method{summary}{simulation}(p)
}
\arguments{
\item{p}{a matrix of class "simulation".}
}
\value{
It returns a summary. The summary contains the number and the length of the simulations and what kind of 
simulation it is. It also contains the value of the parameter phi and the initial value Y_0.
}
\description{
Gives an summary of a simulation.
}
\details{

}
\examples{
p <- rAR1(0, 300, 0.8, distfun = rnorm, 300)
summary(p)
 
}
