% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLN.R
\name{LLN}
\alias{LLN}
\title{LLN function}
\usage{
LLN(y)
}
\arguments{
\item{y}{vector of integers}
}
\description{
Calculates form the first to the the n'th step of sum(Yi)/n, to determine if there is a Law of Large Numbers.
}
\examples{
y<-simAR1(0,1000,0.8,rnorm(1000))
LLN(y)
}
