% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLN.R
\name{LLN}
\alias{LLN}
\title{LLN function}
\usage{
LLN(x)
}
\arguments{
\item{x}{a matrix where every column is a vector of random variables with the same parametervalues.}
}
\value{
A plot through the values \eqn{sum_{i=1}^{k}y_j/k} for each k from 1 to n.
}
\description{
A function that checks, if there is a Law of Large Numbers
Calculates form the first to the the n'th step of sum(Yi)/n, to determine if there is a Law of Large Numbers.
}
\details{
If a given output of this function resembles that it converges towards the same value, it would seem that there is a Law of Large Numbers.
}
\examples{
x<-replicate(10, simAR1(0,1000,0.8,rnorm,1000)
LLN(x)
}
