% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.simulation}
\alias{summary.simulation}
\title{Object summary}
\usage{
\method{summary}{simulation}(p)
}
\arguments{
\item{p}{a matrix of class "simulation".}
}
\value{
It returns a summary. The summary contains the length of the individual simulation, the value of the
parameter phi, the value of the initial value Y_0, and (...).
}
\description{
Gives an summary of a simulation.
}
\details{

}
\examples{
p <- rAR1(0, 300, 0.8, distfun = rnorm, 300)
summary(p)
 
}
